pyproject.toml
[project]
name = "mcvotd"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "geeshta", email = "stefan.foldesi@outlook.com" }
]
dependencies = [
    "aio-mc-rcon>=3.2.2",
    "httpx>=0.27.0",
    "anyio>=4.4.0",
    "apscheduler>=3.10.4",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.5.4",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mcvotd"]
